+++ FullCore/GetParam.h
@@ -65,7 +65,7 @@ template<typename K>
 void FromNumericalVectorToString(std::string & _ai, std::vector<K> mapk, char _separator, int start_pos, int end_pos)
 {
 	if (end_pos == -1)
-		end_pos = mapk.size();
+		end_pos = static_cast<int>(mapk.size());
 
 	for (int i = start_pos; i < end_pos; i++) {
 		_ai += (std::to_string(mapk[i]));
